antarctic-static-site/requirements.txt
Jinja2==3.1.2
Markdown==3.4.3
PyYAML==6.0
python-slugify==8.0.1
Pillow==9.5.0
pygments==2.15.1